Factor out spherical
With usual local imports on top, and import from spherical at bottom
of __init__.py. Also change to import in cyclindrical, and adjustment
of one test, since the location of the classes now has changed.

Not yet changing the locations in the failing doctests.
